requirements.txt
# Core LangChain split packages (modern layout)
langchain>=0.2.14
langchain-core>=0.2.34
langchain-text-splitters>=0.2.2
langchain-chroma>=0.1.1
langchain-ollama>=0.1.0
langchain-huggingface>=0.1.0


# Vector DB
chromadb>=0.5.5


# Loaders / I/O
langchain-community>=0.2.12
pypdf>=4.2.0
python-dotenv>=1.0.1


# Embeddings
sentence-transformers>=3.0.1
huggingface-hub>=0.24.6


# UI
streamlit>=1.37.0


# Optional niceties
tqdm>=4.66.4